refactor!: there can be only one (parser)
