+++ playa/pdftypes.py
@@ -6,11 +6,14 @@ from typing import (
     TYPE_CHECKING,
     Any,
     Dict,
+    Generic,
     Iterable,
     List,
     Optional,
     Protocol,
     Tuple,
+    Type,
+    TypeVar,
     Union,
     cast,
 )
@@ -23,9 +26,9 @@ from playa.exceptions import (
     PDFNotImplementedError,
     PDFTypeError,
     PDFValueError,
+    PSTypeError,
 )
 from playa.lzw import lzwdecode
-from playa.psparser import LIT
 from playa.runlength import rldecode
 from playa.utils import apply_png_predictor
 
@@ -34,8 +37,74 @@ if TYPE_CHECKING:
 
 logger = logging.getLogger(__name__)
 
-LITERAL_CRYPT = LIT("Crypt")
 
+class PSLiteral:
+    """A class that represents a PostScript literal.
+
+    Postscript literals are used as identifiers, such as
+    variable names, property names and dictionary keys.
+    Literals are case sensitive and denoted by a preceding
+    slash sign (e.g. "/Name")
+
+    Note: Do not create an instance of PSLiteral directly.
+    Always use PSLiteralTable.intern().
+    """
+
+    def __init__(self, name: str) -> None:
+        self.name = name
+
+    def __repr__(self) -> str:
+        return "/%r" % self.name
+
+
+class PSKeyword:
+    """A class that represents a PostScript keyword.
+
+    PostScript keywords are a dozen of predefined words.
+    Commands and directives in PostScript are expressed by keywords.
+    They are also used to denote the content boundaries.
+
+    Note: Do not create an instance of PSKeyword directly.
+    Always use PSKeywordTable.intern().
+    """
+
+    def __init__(self, name: bytes) -> None:
+        self.name = name
+
+    def __repr__(self) -> str:
+        return "/%r" % self.name
+
+
+_SymbolT = TypeVar("_SymbolT", PSLiteral, PSKeyword)
+_NameT = TypeVar("_NameT", str, bytes)
+
+
+class PSSymbolTable(Generic[_SymbolT, _NameT]):
+    """Store globally unique name objects or language keywords."""
+
+    def __init__(self, table_type: Type[_SymbolT], name_type: Type[_NameT]) -> None:
+        self.dict: Dict[_NameT, _SymbolT] = {}
+        self.table_type: Type[_SymbolT] = table_type
+        self.name_type: Type[_NameT] = name_type
+
+    def intern(self, name: _NameT) -> _SymbolT:
+        if not isinstance(name, self.name_type):
+            raise ValueError(f"{self.table_type} can only store {self.name_type}")
+        if name in self.dict:
+            lit = self.dict[name]
+        else:
+            lit = self.table_type(name)  # type: ignore
+        self.dict[name] = lit
+        return lit
+
+
+PSLiteralTable = PSSymbolTable(PSLiteral, str)
+PSKeywordTable = PSSymbolTable(PSKeyword, bytes)
+LIT = PSLiteralTable.intern
+KWD = PSKeywordTable.intern
+
+# Intern a bunch of important literals
+LITERAL_CRYPT = LIT("Crypt")
 # Abbreviation of Filter names in PDF 4.8.6. "Inline Images"
 LITERALS_FLATE_DECODE = (LIT("FlateDecode"), LIT("Fl"))
 LITERALS_LZW_DECODE = (LIT("LZWDecode"), LIT("LZW"))
@@ -46,6 +115,51 @@ LITERALS_CCITTFAX_DECODE = (LIT("CCITTFaxDecode"), LIT("CCF"))
 LITERALS_DCT_DECODE = (LIT("DCTDecode"), LIT("DCT"))
 LITERALS_JBIG2_DECODE = (LIT("JBIG2Decode"),)
 LITERALS_JPX_DECODE = (LIT("JPXDecode"),)
+
+
+def name_str(x: bytes) -> str:
+    """Get the string representation for a name object.
+
+    According to the PDF 1.7 spec (p.18):
+
+    > Ordinarily, the bytes making up the name are never treated as
+    > text to be presented to a human user or to an application
+    > external to a conforming reader. However, occasionally the need
+    > arises to treat a name object as text... In such situations, the
+    > sequence of bytes (after expansion of NUMBER SIGN sequences, if
+    > any) should be interpreted according to UTF-8.
+
+    Accordingly, if they *can* be decoded to UTF-8, then they *will*
+    be, and if not, we will just decode them as ISO-8859-1 since that
+    gives a unique (if possibly nonsensical) value for an 8-bit string.
+    """
+    try:
+        return x.decode("utf-8")
+    except UnicodeDecodeError:
+        return x.decode("iso-8859-1")
+
+
+def literal_name(x: Any) -> str:
+    if not isinstance(x, PSLiteral):
+        if settings.STRICT:
+            raise PSTypeError(f"Literal required: {x!r}")
+        return str(x)
+    else:
+        return x.name
+
+
+def keyword_name(x: Any) -> str:
+    if not isinstance(x, PSKeyword):
+        if settings.STRICT:
+            raise PSTypeError("Keyword required: %r" % x)
+        else:
+            return str(x)
+    else:
+        # PDF keywords are *not* UTF-8 (they aren't ISO-8859-1 either,
+        # but this isn't very important, we just want some
+        # unique representation of 8-bit characters, as above)
+        name = x.name.decode("iso-8859-1")
+    return name
 
 
 class DecipherCallable(Protocol):
